servertoclientpacket enum

+++ Voxalia/ServerGame/NetworkSystem/Connection.cs
@@ -51,7 +51,7 @@ namespace Voxalia.ServerGame.NetworkSystem
             {
                 if (Alive)
                 {
-                    byte id = packet.ID;
+                    byte id = (byte)packet.ID;
                     byte[] data = packet.Data;
                     byte[] fdata = new byte[data.Length + 5];
                     Utilities.IntToBytes(data.Length).CopyTo(fdata, 0);
